Can toggle between 2D and 3D views
But it would be nice to keep the camera the same when you toggle between
the views. We could have it so that the 3D figure is generated when
making the annotations, but just not shown, and the button literally
hides / shows the desired graph. Another strategy that is proving more
difficult is the one where the 3D is generated only when you toggle and
only when the found partitions haven't changed, but this is hard because
we would like to be able to store somewhere in one callback that the 3D
figure needs to be updated, and then in another callback that updates
the 3D figure say that it has been updated, but we can't do this. So
maybe there's another way.