It can preserve the 3D scene
But for some reason while you are waiting for the 3D scene to load (when
you toggled to it from the 2D scene) it shows the old 3D scene, but only
when it has loaded does it show the latest 3D scene.

Also this implementation is a little annoying because if you haven't
changed the annotations in 2D, loading in 3D still takes some time
because the 3D figure is still regenerated.