Getting ready for dash-sample-apps
copy_to_dash-sample-apps is so that we can easily keep this repo up to
do date with the folder in dash-sample-apps because it currently seems
that apps cannot be added as submodules (however it seems submodules
within the individual app folders do work...).

Removed test and formatted files in plotly-common.